DeviceGenerator: add EC device and SMC key generation
EC devices do not use deviceId so we generate handlers for matching 'name'.
For each Sensor, we add the 'ReadValue' key to generate inline code for
reading the sensor (instead of calling a parent class function).
We also add 'SmcKey' and 'SmcKeyType' for each sensor to generate code for
SMC key reading.